réarragenemtn des fichiers dans des dossiers et gestions des feux manuels
